Removed custom checks for AWS auth data
There were some checks for environment variables in the Swf handler
class, which causes anything using the class to fatal error out, even if
valid credentials are supplied to the client via another method.

Removed the credentials check, as the aws-php-sdk library will check for
credentials on its own anyway.